compilation error
(void *) is being passed to a function that expects an integer type (ULONGLONG, which is unsigned long). To resolve this issue, you need to cast the pointer to the appropriate type before passing it to the function

+++ dlls/ntdll/unix/loader.c
@@ -1135,7 +1135,7 @@ static NTSTATUS steamclient_setup_trampolines_impl( HMODULE src_mod, HMODULE tgt
         steamclient_tgts[steamclient_count] = tgt_addr;
         if (!noexec_cached)
         {
-            if (x64) steamclient_write_jump_x64(src_addr, tgt_addr);
+            if (x64) steamclient_write_jump_x64(src_addr, (ULONGLONG)tgt_addr);
             else steamclient_write_jump_x86(src_addr, PtrToUlong(tgt_addr));
         }
         else steamclient_count++;
@@ -1161,7 +1161,7 @@ static NTSTATUS steamclient_setup_trampolines_impl( HMODULE src_mod, HMODULE tgt
     steamclient_tgts[steamclient_count] = tgt_addr;
     if (!noexec_cached)
     {
-        if (x64) steamclient_write_jump_x64(src_addr, tgt_addr);
+        if (x64) steamclient_write_jump_x64(src_addr, (ULONGLONG)tgt_addr);
         else steamclient_write_jump_x86(src_addr, PtrToUlong(tgt_addr));
     }
     else steamclient_count++;
